pull duplicate argument cleansing functions into a util file
